refactor(behavior_path_planner): traffic light utils
